KNL-1554 Refactor of time service to aid deprecation
This splits out a new API called UserTimeService which is just responsible for getting the user’s timezone at the moment. This allows the main TimeService to be deprecated and any place that uses TimeService to just get the user’s timezone can switch to using UserTimeService.

This also makes the current TimeService more testable as it doesn’t actually need the whole kernel to be running, just mock instances of 2 helpers that can be easily mocked with Mockito (or similar).

This also makes the classes more focused (Single Responsibility Principal).
