+++ kernel/kernel-impl/src/main/webapp/WEB-INF/components.xml
@@ -20,5 +20,6 @@
      <import resource="event-components.xml" />
      <import resource="memory-components.xml" />
      <import resource="util-components.xml" />
+     <import resource="time-components.xml" />
      <import resource="messagebundle-components.xml" />
 </beans>
